backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
google-genai
sqlalchemy
httpx
python-jose[cryptography]